Fixed bug with IG carousel
