+++ bloom-and-vine/src/Components/instagram-posts/components/InstagramPost.css
@@ -1,8 +1,9 @@
 .curPost {
     position: relative;
-    margin: 0 0.25% 0 0.25%;
     width: 100%;
+    
     aspect-ratio: 1;
+    margin: 0 0.25% 0 0.25%;
 }
 
 .postImg {
@@ -10,8 +11,9 @@
     width: 100%;
     height: 100%;
 
-    object-fit: cover;
     aspect-ratio: 1;
+    object-fit: cover;
+    transition: ease-in 0.4s;
 }
 
 .postImg:hover {
